+++ app/members/tests/__init__.py
@@ -1,0 +1,2 @@
+from .email_login import *
+from .signup import *
